move d3 imports
