+++ d3/axis.ts
@@ -1,0 +1,1 @@
+export { axisLeft } from "https://cdn.skypack.dev/d3-axis@3?dts";
