+++ d3/drag.ts
@@ -1,0 +1,2 @@
+export type { DragBehavior } from "https://cdn.skypack.dev/d3-drag@3?dts";
+export { drag } from "https://cdn.skypack.dev/d3-drag@3?dts";
