+++ d3/force.ts
@@ -1,0 +1,15 @@
+export type {
+  ForceLink,
+  Simulation,
+  SimulationLinkDatum,
+  SimulationNodeDatum,
+} from "https://cdn.skypack.dev/d3-force@3?dts";
+export {
+  forceCenter,
+  forceCollide,
+  forceLink,
+  forceManyBody,
+  forceSimulation,
+  forceX,
+  forceY,
+} from "https://cdn.skypack.dev/d3-force@3?dts";
