+++ d3/scale.ts
@@ -1,0 +1,5 @@
+export {
+  scaleLinear,
+  scaleOrdinal,
+  scaleTime,
+} from "https://cdn.skypack.dev/d3-scale@4?dts";
