+++ d3/shape.ts
@@ -1,0 +1,1 @@
+export { line } from "https://cdn.skypack.dev/d3-shape@3?dts";
