--- visualisation/d3.ts
@@ -1,30 +1,0 @@
-export { create } from "https://cdn.skypack.dev/d3-selection@3?dts";
-
-// D3 Force
-export type {
-  ForceLink,
-  Simulation,
-  SimulationLinkDatum,
-  SimulationNodeDatum,
-} from "https://cdn.skypack.dev/d3-force@3?dts";
-export { axisLeft } from "https://cdn.skypack.dev/d3-axis@3?dts";
-export {
-  scaleLinear,
-  scaleOrdinal,
-  scaleTime,
-} from "https://cdn.skypack.dev/d3-scale@4?dts";
-export { line } from "https://cdn.skypack.dev/d3-shape@3?dts";
-
-export {
-  // forceCenter,
-  forceCollide,
-  forceLink,
-  forceManyBody,
-  forceSimulation,
-  forceX,
-  forceY,
-} from "https://cdn.skypack.dev/d3-force@3?dts";
-
-// D3 Drag
-export type { DragBehavior } from "https://cdn.skypack.dev/d3-drag@3?dts";
-export { drag } from "https://cdn.skypack.dev/d3-drag@3?dts";
